Fixed bug in parsing names with tildes.
Fixed a bug that caused names containing tildes to be parsed
incorrectly. Within a group expression of a regex, the hyphen character
takes on special meaning when between two other characters. In such a
case, it means use a range of characters (like, a-z). Instead, we just
want to match the hyphen character. So, move the hyphen to the end of
the group expression and add the tilde right before it.